refactor: インスタンス化reflection廃止
