+++ iplass-gem/src/main/java/org/iplass/gem/command/generic/search/FixedSearchCommand.java
@@ -20,6 +20,7 @@
 
 package org.iplass.gem.command.generic.search;
 
+import org.iplass.mtp.command.RequestContext;
 import org.iplass.mtp.command.annotation.CommandClass;
 
 @CommandClass(name = FixedSearchCommand.CMD_NAME, displayName = "定型検索")
@@ -28,7 +29,12 @@ public final class FixedSearchCommand extends SearchCommandBase {
 	public static final String CMD_NAME = "gem/generic/search/FixedSearchCommand";
 
 	@Override
-	protected Class<? extends SearchContext> getContextClass() {
-		return FixedSearchContext.class;
+	public SearchContext getContext(RequestContext request) {
+		SearchContext context = new FixedSearchContext();
+		context.setRequest(request);
+		context.setEntityDefinition(edm.get(context.getDefName()));
+		context.setEntityView(evm.get(context.getDefName()));
+
+		return context;
 	}
 }
